break subtitle into multiple files + implement subtitle visuals
